Fix case of $0 truncating text due to being frozen
Fix failure on test "sets $0 while working"
This appears to be an OS dependent issue. I ran into this on
Ubuntu Linux 10.04 64bit on kernel 2.6.32-22
I do not know if this is an ideal solution, not did I come up
with it. I found the solution at:
http://www.ruby-forum.com/topic/187083#817010
but I do know that the tests now pass, so hopefully either this
gets incorporated, or triggers the creation of a better solution.